+++ src/NonEmptyContainers/SeqConversions.hs
@@ -1,3 +1,5 @@
+{-# LANGUAGE FlexibleInstances #-}
+{-# LANGUAGE InstanceSigs      #-}
 {-# LANGUAGE NoImplicitPrelude #-}
 {-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE TypeFamilies      #-}
@@ -11,9 +13,12 @@ where
 
 -- base --------------------------------
 
+import qualified  Data.Foldable
+
 import Control.Applicative  ( pure )
 import Data.Eq              ( Eq )
 import Data.Function        ( ($), id )
+import Data.List.NonEmpty   ( NonEmpty, toList )
 import Data.Maybe           ( Maybe( Just, Nothing ) )
 import Data.String          ( String )
 import System.Exit          ( ExitCode )
@@ -32,12 +37,13 @@ import Data.Sequence  ( Seq( (:<|) ) )
 
 -- lens --------------------------------
 
-import Control.Lens.Iso    ( Iso' )
+import Control.Lens.Iso    ( Iso', iso )
 import Control.Lens.Prism  ( Prism', prism' )
 
 -- mono-traversable --------------------
 
 import Data.MonoTraversable  ( Element )
+import Data.NonNull          ( NonNull, toNullable )
 
 -- more-unicode ------------------------
 
@@ -77,6 +83,16 @@ class FromMonoSeq α where
   fromList ∷ [Element α] → α
   fromList = fromSeq ∘ Seq.fromList
 
+instance FromMonoSeq (Seq α) where
+  fromSeq = id
+  fromList ∷ [α] → Seq α
+  fromList = Seq.fromList
+
+instance FromMonoSeq [α] where
+  fromSeq ∷ Seq α → [α]
+  fromSeq = Data.Foldable.toList
+  fromList = id
+
 ----------------------------------------
 --             ToMonoSeq              --
 ----------------------------------------
@@ -85,11 +101,20 @@ class FromMonoSeq α where
 class ToMonoSeq α where
   toSeq ∷ α → Seq (Element α)
 
-instance α ~ Element (SeqNE α) ⇒ ToMonoSeq (SeqNE α) where
+instance {- α ~ Element (Seq α) ⇒ -} ToMonoSeq (Seq α) where
+  toSeq = id
+
+instance {- α ~ Element (SeqNE α) ⇒ -} ToMonoSeq (SeqNE α) where
   toSeq = SeqNE.toSeq
 
-instance α ~ Element (Seq α) ⇒ ToMonoSeq (Seq α) where
-  toSeq = id
+instance ToMonoSeq [α] where
+  toSeq = Seq.fromList
+
+instance ToMonoSeq (NonEmpty α) where
+  toSeq = toSeq ∘ toList
+
+instance ToMonoSeq (NonNull [α]) where
+  toSeq = toSeq ∘ toNullable
 
 ----------------------------------------
 --             IsMonoSeq              --
@@ -99,16 +124,34 @@ instance α ~ Element (Seq α) ⇒ ToMonoSeq (Seq α) where
 class IsMonoSeq α where
   seq ∷ Iso' α (Seq (Element α))
 
+instance IsMonoSeq (Seq α) where
+  seq = id
+
+instance IsMonoSeq [α] where
+  seq = iso toSeq fromSeq
+
 ----------------------------------------
 --             AsMonoSeq              --
 ----------------------------------------
 
-{- | α that may be representable by a non-empty Sequence of `Element α` -}
-class AsMonoSeq α where
+{- | α that may be representable by a Sequence of `Element α`; i.e., a
+     `Seq (Element α)` is always convertable to an α, and an α /may be/
+     convertable to a `Seq (Element α)`.  So there must be a `FromMonoSeq`
+     instance.
+ -}
+class FromMonoSeq α ⇒ AsMonoSeq α where
+  toSeqMay ∷ α → Maybe (Seq (Element α))
   seq' ∷ Prism' α (Seq (Element α))
+  seq' = prism' fromSeq toSeqMay
 
-instance α ~ Element (Seq α) ⇒ AsMonoSeq (Seq α) where
-  seq' = prism' id Just
+instance {- α ~ Element (Seq α) ⇒ -} AsMonoSeq (Seq α) where
+--  seq' = id -- prism' id Just
+  toSeqMay = Just
+
+instance AsMonoSeq [α] where
+  seq' = -- prism' (Seq α → [α]) ([α] → Maybe (Seq α))
+         prism' fromSeq (Just ∘ toSeq)
+  toSeqMay = Just ∘ toSeq
 
 ------------------------------------------------------------